requirements.txt
absl-py==2.2.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.45.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.71.0
h11==0.16.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.3.1
multiprocess==0.70.16
narwhals==1.36.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.8.0.87
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
openai==1.76.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.1
pexpect==4.9.0
pillow==11.2.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-triton==3.3.0+git96316ce5
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==78.1.0
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tokenizers==0.21.1
torch==2.8.0.dev20250426+cu128
torchaudio==2.6.0.dev20250426+cu128
torchvision==0.22.0.dev20250426+cu128
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.2.0
trl==0.15.2
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tyro==0.9.18
tzdata==2025.2
unsloth==2025.3.19
unsloth_zoo==2025.3.17
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
xformers==0.0.30+836cd905.d20250405  # from source
